// src/models/ITrackingDocument.ts
export interface ITrackingDocument {
  success: boolean;
  data: [
    {
      LightReturnNumber: string;
      CounterpartyRecipientDescription: string;
      DocumentWeight: number;
      ServiceType: string;
      UndeliveryReasons: [];
      LoyaltyCardSender: string;
      RecipientFullName: string;
      FactualWeight: string;
      MarketplacePartnerToken: string;
      CounterpartySenderDescription: string;
      InternationalDeliveryType: string;
      CargoType: string;
      PayerType: string;
      SeatsAmount: string;
      ScheduledDeliveryDate: string;
      DocumentCost: string;
      CardMaskedNumber: string;
      OwnerDocumentType: string;
      ExpressWaybillPaymentStatus: string;
      ExpressWaybillAmountToPay: string;
      AfterpaymentOnGoodsCost: string;
      SumBeforeCheckWeight: number;
      CheckWeight: number;
      PaymentMethod: string;
      AdjustedDate: string;
      Number: string;
      TrackingUpdateDate: string;
      CalculatedWeight: string;
      WarehouseRecipient: string;
      WarehouseSender: string;
      DateCreated: string;
      DateScan: string;
      DateMoving: string;
      DateFirstDayStorage: string;
      DatePayedKeeping: string;
      RecipientAddress: string;
      RecipientDateTime: string;
      RefCityRecipient: string;
      RefCitySender: string;
      RefSettlementRecipient: string;
      RefSettlementSender: string;
      SenderAddress: string;
      ClientBarcode: string;
      CitySender: string;
      CityRecipient: string;
      CargoDescriptionString: string;
      AnnouncedPrice: string;
      AdditionalInformationEW: string;
      ActualDeliveryDate: string;
      StatusCode: string;
      PostomatV3CellReservationNumber: boolean;
      AmountToPay: string;
      AmountPaid: string;
      RefEW: string;
      VolumeWeight: string;
      CheckWeightMethod: string;
      OwnerDocumentNumber: string;
      LastCreatedOnTheBasisNumber: string;
      LastCreatedOnTheBasisDateTime: string;
      LastTransactionDateTimeGM: string;
      PaymentStatus: string;
      PaymentStatusDate: string;
      LastAmountTransferGM: string;
      LastAmountReceivedCommissionGM: number;
      LastCreatedOnTheBasisPayerType: string;
      DeliveryTimeframe: string;
      LastTransactionStatusGM: string;
      Status: string;
      CreatedOnTheBasis: string;
      Redelivery: number;
      RedeliveryNum: string;
      RedeliverySum: string;
      RedeliveryPayer: string;
      RecipientWarehouseTypeRef: string;
      WarehouseRecipientInternetAddressRef: string;
      WarehouseRecipientNumber: number;
      WarehouseRecipientRef: string;
      CategoryOfWarehouse: string;
      WarehouseRecipientAddress: string;
      WarehouseSenderInternetAddressRef: string;
      WarehouseSenderAddress: string;
      CounterpartyType: string;
      CounterpartySenderType: string;
      AviaDelivery: number;
      BarcodeRedBox: string;
      CargoReturnRefusal: boolean;
      DaysStorageCargo: string;
      Packaging: [];
      PartialReturnGoods: [];
      SecurePayment: boolean;
      StorageAmount: string;
      StoragePrice: string;
      PossibilityCreateRedirecting: false;
      PossibilityCreateReturn: false;
      PossibilityCreateRefusal: false;
      PossibilityChangeEW: false;
      PossibilityChangeCash2Card: false;
      PossibilityChangeDeliveryIntervals: false;
      PossibilityTrusteeRecipient: false;
      TrusteeRecipientPhone: string;
      PossibilityLightReturn: true;
      RedeliveryPaymentCardRef: string;
      RedeliveryPaymentCardDescription: string;
      FreeShipping: string;
      InternetDocumentDescription: string;
      LastCreatedOnTheBasisDocumentType: string;
      LoyaltyCardRecipient: string;
      PhoneRecipient: string;
      PhoneSender: string;
      RecipientFullNameEW: string;
      RedeliveryServiceCost: string;
      SenderFullNameEW: string;
      UndeliveryReasonsDate: string;
      UndeliveryReasonsSubtypeDescription: string;
    }
  ];
  errors: [];
  warnings: [
    {
      ID_20450749133426: string;
    }
  ];
  info: [];
  messageCodes: [];
  errorCodes: [];
  warningCodes: [];
  infoCodes: [];
}